Conflicts:
	panda/src/linmath/lvecBase4_src.I

+++ direct/src/plugin_npapi/ppBrowserObject.cxx
@@ -230,10 +230,10 @@ call(const string &method_name, P3D_object *params[], int num_params) const {
 ////////////////////////////////////////////////////////////////////
 P3D_object *PPBrowserObject::
 eval(const string &expression) const {
-  NPString npexpr = { expression.c_str(), expression.length() };
+  NPString npexpr = { expression.c_str(), (uint32_t)expression.length() };
 
   NPVariant result;
-  if (!browser->evaluate(_instance->get_npp_instance(), _npobj, 
+  if (!browser->evaluate(_instance->get_npp_instance(), _npobj,
                          &npexpr, &result)) {
     // Failed to eval.
     return NULL;
