+++ panda/src/pnmimage/pnm-image-filter-sparse-core.cxx
@@ -19,7 +19,7 @@
 
 static void
 FUNCTION_NAME(IMAGETYPE &dest, const IMAGETYPE &source,
-              double width, FilterFunction *make_filter, int channel) {
+              float width, FilterFunction *make_filter, int channel) {
   if (!dest.is_valid() || !source.is_valid()) {
     return;
   }
@@ -40,17 +40,17 @@ FUNCTION_NAME(IMAGETYPE &dest, const IMAGETYPE &source,
   }
 
   // First, scale the image in the A direction.
-  double scale;
+  float scale;
   StoreType *temp_source, *temp_source_weight, *temp_dest, *temp_dest_weight;
 
-  scale = (double)dest.ASIZE() / (double)source.ASIZE();
+  scale = (float)dest.ASIZE() / (float)source.ASIZE();
   temp_source = (StoreType *)PANDA_MALLOC_ARRAY(source.ASIZE() * sizeof(StoreType));
   temp_source_weight = (StoreType *)PANDA_MALLOC_ARRAY(source.ASIZE() * sizeof(StoreType));
   temp_dest = (StoreType *)PANDA_MALLOC_ARRAY(dest.ASIZE() * sizeof(StoreType));
   temp_dest_weight = (StoreType *)PANDA_MALLOC_ARRAY(dest.ASIZE() * sizeof(StoreType));
 
   WorkType *filter;
-  double filter_width;
+  float filter_width;
 
   make_filter(scale, width, filter, filter_width);
 
@@ -81,7 +81,7 @@ FUNCTION_NAME(IMAGETYPE &dest, const IMAGETYPE &source,
   PANDA_FREE_ARRAY(filter);
 
   // Now, scale the image in the B direction.
-  scale = (double)dest.BSIZE() / (double)source.BSIZE();
+  scale = (float)dest.BSIZE() / (float)source.BSIZE();
   temp_dest = (StoreType *)PANDA_MALLOC_ARRAY(dest.BSIZE() * sizeof(StoreType));
   temp_dest_weight = (StoreType *)PANDA_MALLOC_ARRAY(dest.BSIZE() * sizeof(StoreType));
 
@@ -95,7 +95,7 @@ FUNCTION_NAME(IMAGETYPE &dest, const IMAGETYPE &source,
 
     for (b = 0; b < dest.BSIZE(); b++) {
       if (temp_dest_weight[b] != 0) {
-        dest.SETVAL(a, b, channel, (double)temp_dest[b]/(double)source_max);
+        dest.SETVAL(a, b, channel, (float)temp_dest[b]/(float)source_max);
       }
     }
   }
